Added capacitor build
Add to AndroidManifest.xml.  Not sure if this is needed:

android:usesCleartextTraffic="true"
android:networkSecurityConfig="@xml/network_security_config"

Added "cleartext" to capacitor.config.json and ran "npx cap copy", which
fixed the "Failed to fetch" error.

Need to use the IP address (not localhost) for HASURA_GRAPHQL_URL,
because Android uses localhost within the emulator (it won't reach the
service on the host).